feat: fetching framework functional
